Extract a separate testCommon module
