+++ backendImpl/build.gradle
@@ -11,6 +11,7 @@ dependencies {
   api project(':backendApi')
   api project(':branchLayoutApi')
 
+  testImplementation project(':testCommon').sourceSets.test.output
   testRuntime project(':branchLayoutImpl')
   testRuntime project(':gitCoreJGit')
 }
