+++ backendImpl/src/test/java/com/virtuslab/gitmachete/backend/integration/GitMacheteStatusTest.java
@@ -19,15 +19,15 @@ import org.junit.Test;
 import com.virtuslab.binding.RuntimeBinding;
 import com.virtuslab.branchlayout.api.IBranchLayout;
 import com.virtuslab.branchlayout.api.manager.IBranchLayoutReader;
-import com.virtuslab.gitmachete.backend.BaseGitRepositoryTest;
 import com.virtuslab.gitmachete.backend.api.BaseGitMacheteBranch;
 import com.virtuslab.gitmachete.backend.api.BaseGitMacheteNonRootBranch;
 import com.virtuslab.gitmachete.backend.api.BaseGitMacheteRootBranch;
 import com.virtuslab.gitmachete.backend.api.IGitMacheteRepository;
 import com.virtuslab.gitmachete.backend.api.SyncToParentStatus;
 import com.virtuslab.gitmachete.backend.impl.GitMacheteRepositoryFactory;
+import com.virtuslab.gitmachete.testcommon.BaseGitRepositoryBackedTest;
 
-public class GitMacheteStatusTest extends BaseGitRepositoryTest {
+public class GitMacheteStatusTest extends BaseGitRepositoryBackedTest {
   IGitMacheteRepository gitMacheteRepository = null;
   GitMacheteRepositoryFactory gitMacheteRepositoryFactory = new GitMacheteRepositoryFactory();
   IBranchLayoutReader branchLayoutReader = RuntimeBinding.instantiateSoleImplementingClass(IBranchLayoutReader.class);
