+++ settings.gradle
@@ -2,6 +2,7 @@ rootProject.name = 'git-machete-intellij-plugin'
 include 'logging'
 include 'binding'
 include 'qual'
+include 'testCommon'
 include 'gitCoreApi'
 include 'gitCoreJGit'
 include 'branchLayoutApi'
