+++ testCommon/build.gradle
@@ -1,0 +1,1 @@
+junit()
