+++ testCommon/src/test/resources/common.sh
