+++ testCommon/src/test/resources/setup-for-diverged-and-older-than.sh
