+++ testCommon/src/test/resources/setup-with-single-remote.sh
